requested API changes for coinmarketcap support. (#46)
* added supply with support for future epocs
* refactoring to reduce database calls

+++ src/Akroma.Domain/NetworkStats/Models/Stats.cs
@@ -8,6 +8,6 @@ namespace Akroma.Domain.NetworkStats.Models
         public string Difficulty { get; set; }
         public string HashRate { get; set; }
         public DateTime CreatedAt { get; set; }
-        public int CirculatingSupply { get; set; }
+        public double CirculatingSupply { get; set; }
     }
 }
